refactor(snippets): move WordsSource
